Boring Changes and An HTML File.
For the index page, I created an HTML page, so that you can see the
edits you make real time. Additionally, I added, in the comments, the
basics of the require stuff we’re doing with php, as well as a few—and
I mean few—CSS basics. This is more important to the PHP segment, to
the commit goes into the PHP branch.